refactor: reorganize processors by data type instead of technical concerns
- Group health-related files under processors/health/
- Group training-related files under processors/training/
- Group macros-related files under processors/macros/
- Mirror same structure in tests/processors/
- Update all import paths to reflect new structure
- Add readonly modifiers to private fields that don't change
- Use nullish coalescing and optional chaining for cleaner code
- Maintain all existing functionality and test coverage
