+++ processors/health/constants.ts
