+++ processors/training/summary.ts
